no limit implemented, without ortho

+++ include/filters.h
@@ -3,9 +3,9 @@
 
 #include "nn_table.h"
 
-bool gc_check(seed *s, uint64_t base, int max_gc_count);
+bool gc_check(seed *s, uint8_t base, int max_gc_count);
 
-bool hp_check(seed *s, uint64_t base, int hp_max);
+bool hp_check(seed *s, uint8_t base, int hp_max);
 
 int hamming_check(uint64_t a, uint64_t b);
